telethon_user/src/requirements.txt
telethon==1.24
teleredis==0.1.3
aioredis==2.0.0
croniter==1.0.15
pydantic==1.10.2
